move test related packages under common folder
