--- deputy-claimer/integration_test/deputy/btcb-config.json
@@ -1,88 +1,0 @@
-{
-  "db_config": {
-    "dialect": "sqlite3",
-    "db_path": "/deputy/deputy_btcb.db",
-    "max_bnb_kept_block_height": 200000,
-    "max_other_kept_block_height": 200000
-  },
-  "alert_config": {
-    "telegram_bot_id": "",
-    "telegram_chat_id": "",
-    "bnb_block_update_time_out": 60,
-    "other_chain_block_update_time_out": 600,
-    "reconciliation_diff_amount": "10"
-  },
-  "chain_config": {
-    "bnb_confirm_num": 2,
-    "bnb_auto_retry_num": 3,
-    "bnb_auto_retry_timeout": 60,
-    "bnb_expire_height_span": 360,
-    "bnb_min_accept_expire_height_span": 10000,
-    "bnb_min_remain_height": 1000,
-    "bnb_min_swap_amount": 3,
-    "bnb_max_swap_amount": 100000000,
-    "bnb_max_deputy_out_amount": 100000000,
-    "bnb_ratio": "1",
-    "bnb_fixed_fee": 2,
-    "bnb_start_height": 1,
-    "bnb_hot_wallet_overflow": 10000000000,
-    "other_chain": "KAVA",
-    "other_chain_confirm_num": 2,
-    "other_chain_decimal": 8,
-    "other_chain_expire_height_span": 220,
-    "other_chain_auto_retry_num": 3,
-    "other_chain_auto_retry_timeout": 1200,
-    "other_chain_min_accept_expire_height_span": 200,
-    "other_chain_min_remain_height": 200,
-    "other_chain_min_swap_amount": 3,
-    "other_chain_max_swap_amount": 100000000,
-    "other_chain_max_deputy_out_amount": 100000000,
-    "other_chain_ratio": "1",
-    "other_chain_fixed_fee": 2,
-    "other_chain_start_height": 1,
-    "other_chain_hot_wallet_overflow": 10000000000
-  },
-  "log_config": {
-    "level": "INFO",
-    "filename": "",
-    "max_file_size_in_mb": 0,
-    "max_backups_of_log_files": 0,
-    "max_age_to_retain_log_files_in_days": 0,
-    "use_console_logger": true,
-    "use_file_logger": false,
-    "compress": false
-  },
-  "admin_config": {
-    "listen_addr": "127.0.0.1:8081"
-  },
-  "instrumentation_config": {
-    "prometheus": true,
-    "prometheus_listen_addr": "127.0.0.1:9090"
-  },
-  "bnb_config": {
-    "key_type": "mnemonic",
-    "aws_region": "",
-    "aws_secret_name": "",
-    "mnemonic": "enjoy soldier replace ugly glimpse rude sponsor hood jewel inner hole tower initial drive jungle resist answer display capable give lesson mule spray whisper",
-    "rpc_addr": "http://bnbnode:26657",
-    "symbol": "BTCB-1DE",
-    "deputy_addr": "bnb1z8ryd66lhc4d9c0mmxx9zyyq4t3cqht9mt0qz3",
-    "cold_wallet_addr": "bnb1tpgqfslnm486qtnfatewdeya2khaav3x6hqhf9",
-    "fetch_interval": 2,
-    "token_balance_alert_threshold": 10000,
-    "bnb_balance_alert_threshold": 10000
-  },
-  "kava_config": {
-    "key_type": "mnemonic",
-    "aws_region": "",
-    "aws_secret_name": "",
-    "mnemonic": "shed crush identify inmate fault truck raw sausage afford fiction day delay people shrimp firm group maple square host thank motor radio visual cable",
-    "rpc_addr": "tcp://kavanode:26657",
-    "symbol": "btcb",
-    "deputy_addr": "kava1kla4wl0ccv7u85cemvs3y987hqk0afcv7vue84",
-    "cold_wallet_addr": "kava1ynf22ap74j6znl503a56y23x5stfr0aw5kntp8",
-    "fetch_interval": 2,
-    "token_balance_alert_threshold": 10000,
-    "kava_balance_alert_threshold": 10000
-  }
-}
