--- deputy-claimer/integration_test/deputy/docker-compose.yaml
@@ -1,18 +1,0 @@
-version: "3"
-services:
-  bnb_deputy:
-    image: kava/deputy:46d9406
-    volumes:
-      - "./deputy/bnb-config.json:/deputy/config/config.json"
-  btcb_deputy:
-    image: kava/deputy:46d9406
-    volumes:
-      - "./deputy/btcb-config.json:/deputy/config/config.json"
-  busd_deputy:
-    image: kava/deputy:46d9406
-    volumes:
-      - "./deputy/busd-config.json:/deputy/config/config.json"
-  xrpb_deputy:
-    image: kava/deputy:46d9406
-    volumes:
-      - "./deputy/xrpb-config.json:/deputy/config/config.json"
