--- deputy-claimer/integration_test/deputy/xrpb-config.json
@@ -1,88 +1,0 @@
-{
-  "db_config": {
-    "dialect": "sqlite3",
-    "db_path": "/deputy/deputy_xrpb.db",
-    "max_bnb_kept_block_height": 200000,
-    "max_other_kept_block_height": 200000
-  },
-  "alert_config": {
-    "telegram_bot_id": "",
-    "telegram_chat_id": "",
-    "bnb_block_update_time_out": 60,
-    "other_chain_block_update_time_out": 600,
-    "reconciliation_diff_amount": "10"
-  },
-  "chain_config": {
-    "bnb_confirm_num": 2,
-    "bnb_auto_retry_num": 3,
-    "bnb_auto_retry_timeout": 60,
-    "bnb_expire_height_span": 360,
-    "bnb_min_accept_expire_height_span": 10000,
-    "bnb_min_remain_height": 1000,
-    "bnb_min_swap_amount": 0,
-    "bnb_max_swap_amount": 10000000000000,
-    "bnb_max_deputy_out_amount": 10000000000000,
-    "bnb_ratio": "1",
-    "bnb_fixed_fee": 100000,
-    "bnb_start_height": 1,
-    "bnb_hot_wallet_overflow": 100000000000000,
-    "other_chain": "KAVA",
-    "other_chain_confirm_num": 2,
-    "other_chain_decimal": 8,
-    "other_chain_expire_height_span": 220,
-    "other_chain_auto_retry_num": 3,
-    "other_chain_auto_retry_timeout": 1200,
-    "other_chain_min_accept_expire_height_span": 200,
-    "other_chain_min_remain_height": 200,
-    "other_chain_min_swap_amount": 1,
-    "other_chain_max_swap_amount": 10000000000000,
-    "other_chain_max_deputy_out_amount": 10000000000000,
-    "other_chain_ratio": "1",
-    "other_chain_fixed_fee": 100000,
-    "other_chain_start_height": 1,
-    "other_chain_hot_wallet_overflow": 100000000000000
-  },
-  "log_config": {
-    "level": "INFO",
-    "filename": "",
-    "max_file_size_in_mb": 0,
-    "max_backups_of_log_files": 0,
-    "max_age_to_retain_log_files_in_days": 0,
-    "use_console_logger": true,
-    "use_file_logger": false,
-    "compress": false
-  },
-  "admin_config": {
-    "listen_addr": "127.0.0.1:8081"
-  },
-  "instrumentation_config": {
-    "prometheus": true,
-    "prometheus_listen_addr": "127.0.0.1:9090"
-  },
-  "bnb_config": {
-    "key_type": "mnemonic",
-    "aws_region": "",
-    "aws_secret_name": "",
-    "mnemonic": "forward argue march dignity puzzle celery caught maze judge chair cement choice bamboo pulse else local foam abuse crazy bullet feed hero rose seat",
-    "rpc_addr": "http://bnbnode:26657",
-    "symbol": "XRP-BF2",
-    "deputy_addr": "bnb1ryrenacljwghhc5zlnxs3pd86amta3jcaagyt0",
-    "cold_wallet_addr": "bnb13l6w5vzqa3533ukrpzpycqh62qd4r3tman8cnv",
-    "fetch_interval": 2,
-    "token_balance_alert_threshold": 10000,
-    "bnb_balance_alert_threshold": 10000
-  },
-  "kava_config": {
-    "key_type": "mnemonic",
-    "aws_region": "",
-    "aws_secret_name": "",
-    "mnemonic": "trial friend silly sugar maid behave slim onion swap report inmate hold hammer hip wrist above sketch mean fence reason master green panel chimney",
-    "rpc_addr": "tcp://kavanode:26657",
-    "symbol": "xrpb",
-    "deputy_addr": "kava14q5sawxdxtpap5x5sgzj7v4sp3ucncjlpuk3hs",
-    "cold_wallet_addr": "kava1z3ytjpr6ancl8gw80z6f47z9smug7986x29vtj",
-    "fetch_interval": 2,
-    "token_balance_alert_threshold": 10000,
-    "kava_balance_alert_threshold": 10000
-  }
-}
