+++ deputy-claimer/test/integration/binance/initstate/.bnbchaind/config/genesis.json
