+++ deputy-claimer/test/integration/binance/initstate/.bnbchaind/config/init-data-directory.sh
