+++ deputy-claimer/test/integration/binance/initstate/.bnbchaind/config/priv_validator_state.json.example
