+++ deputy-claimer/test/integration/docker-compose.yaml
@@ -1,8 +1,4 @@
 services:
-    # bnb_deputy:
-    #     image: kava/deputy:46d9406
-    #     volumes:
-    #         - ./deputy/bnb-config.json:/deputy/config/config.json
     bnbnode:
         build: ./binance
         command:
@@ -24,14 +20,6 @@ services:
             - 8080:8080
         volumes:
             - ./binance/initstate/.bnbcli:/root/.bnbcli
-    # btcb_deputy:
-    #     image: kava/deputy:46d9406
-    #     volumes:
-    #         - ./deputy/btcb-config.json:/deputy/config/config.json
-    # busd_deputy:
-    #     image: kava/deputy:46d9406
-    #     volumes:
-    #         - ./deputy/busd-config.json:/deputy/config/config.json
     kavanode:
         command:
             - sh
@@ -53,8 +41,4 @@ services:
         image: kava/kava:master
         ports:
             - 1317:1317
-    # xrpb_deputy:
-    #     image: kava/deputy:46d9406
-    #     volumes:
-    #         - ./deputy/xrpb-config.json:/deputy/config/config.json
 version: "3"
