+++ deputy-claimer/test/integration/kava/initstate/.kvd/config/addrbook.json
