+++ deputy-claimer/test/integration/kava/initstate/.kvd/config/genesis.json
