+++ deputy-claimer/test/integration/kava/initstate/.kvd/config/gentx/gentx-23a5601684fa065ce9d953af5fc7d4347050ad14.json
