+++ deputy-claimer/test/integration/kava/initstate/.kvd/config/init-data-directory.sh
