+++ deputy-claimer/test/integration/kava/initstate/.kvd/config/priv_validator_state.json.example
