+++ deputy-claimer/test/swap/swaps.go
@@ -1,4 +1,4 @@
-package integrationtest
+package swap
 
 import (
 	"fmt"
